[REFACTORING] reducing the MEW io complexity, such as splitting classes by modules; adding the refactoring required for the latest changes in MEW models
